chore: refactor worker
