+++ apps/api/scripts/generate-api-key.ts
@@ -33,7 +33,7 @@ try {
 
   // write to .env file
   fs.appendFileSync(".env", `TESTING_API_KEY=${key}\n`);
-  fs.appendFileSync("../video-worker/.env", `\nTESTING_API_KEY=${key}\n`);
+  fs.appendFileSync("../video-worker/.env", `\nAPI_KEY=${key}\n`);
   fs.appendFileSync(".env", `\nTESTING_API_KEY=${key}`);
   fs.appendFileSync("../web/.env", `\nAPI_KEY=${key}`);
