+++ apps/api/src/config/file-types.ts
@@ -1,0 +1,4 @@
+export enum FILE_TYPES {
+    VIDEO = "video",
+    TRANSCRIPT = "transcript",
+}
